Titles Listing - Multiselection of filter: quick fix
Because the parameters could be either a List of String or a single String,
there were case scenario where we were using String.contains() instead of List.constains()
and value where matching when they should not (i.e. "56".contains("6")).
It has been fixed by using params.list() to make sure we get a List and not a String.