IssueID #10406 Added potential fix for slow subscription export
These changes have improved the export time on my local machine from 3.30 mins to 12 seconds committing to newdev so we can see if the speed improvements are seen there too.

I have only changed the relevant section of the xml export to use the new version of the method all the other calls to the db version of the method have been left alone.

Looks like the lookupOrCreate methods are the only place that the TIPP Access Status values are added to the database so now I've removed them they will never be added.  I don't
think this is an issue if I'm never going to retrieve them anywhere else and don't need them for the export also another calls to lookupOrCreate will create them.